moved departmentdata, general restructure
