+++ src/components/Nav.js
@@ -1,6 +1,7 @@
 import React from "react";
 import { Link } from "react-router-dom";
 import Logo from "../img/logo.png";
+import "../public"
