+++ src/components/departmentdata.js
@@ -1,5 +1,5 @@
 import React from "react";
-import DepCard from "./components/depcard"
+import ""
 
 const DepartmentData = () => {
     const url = "https://collectionapi.metmuseum.org/public/collection/v1/departments"
@@ -12,14 +12,18 @@ const DepartmentData = () => {
         const getDepartment = async () => {
             const response = await fetch(url);
             const data = await response.json();
-            const dept = data.departments.map((dept) => (dept.displayName))
+            const dept = data.departments.map((dept) => (
+                <div className="departmentcards" key={dept.departmentId}>
+                    <h4>{dept.displayName}</h4>
+                </div>
+                ))
             setDepartment(dept);
             // console.log(dep)
             // setDeptBool(false)
         };
         getDepartment();
     }, []);
-console.log(department)
+    // console.log(department)
     // console.log(department.departments)
 
     // const dep = department.departments.map(dept => console.log(dept))
