--- src/components/depcard.js
@@ -1,11 +1,0 @@
-import DepartmentData from "../departmentdata"
-
-const DepCard = (props) => {
-    return (
-        <div>
-            <h1>{props.data}</h1>
-        </div>
-    )
-}
-
-export default DepCard;
